add utils file to avoid duplication
